feat: add ClientContextProvider component; create hooks directory
