+++ src/hooks/use-client.js
@@ -1,0 +1,4 @@
+import { useContext } from "react";
+import { ClientContext } from "../store";
+
+export const useClient = () => useContext(ClientContext);
