+++ src/hooks/use-current-user.js
@@ -1,4 +1,4 @@
 import { useContext } from "react";
-import { CurrentUserContext } from "./CurrentUserContext";
+import { CurrentUserContext } from "../store";
 
 export const useCurrentUser = () => useContext(CurrentUserContext);
