+++ src/store/index.js
@@ -1,3 +1,6 @@
+export * from "./client-context";
+export * from "./client-context-provider";
+export * from "./current-user-context";
 export * from "./current-user-context-provider";
+export * from "./modal-context";
 export * from "./modal-context-provider";
-export * from "./use-modals";
